+++ apps/vscode-e2e/wdio.conf.ts
@@ -304,8 +304,23 @@ export const config: Options.Testrunner = {
    * @param {Boolean} result.passed    true if test has passed, otherwise false
    * @param {Object}  result.retries   informations to spec related retries, e.g. `{ attempts: 0, limit: 0 }`
    */
-  // afterTest: function(test, context, { error, result, duration, passed, retries }) {
-  // },
+  afterTest: function (
+    test,
+    context,
+    { error, result, duration, passed, retries }
+  ) {
+    if (error) {
+      if (!existsSync('./.screenshots')) {
+        mkdirSync('./.screenshots');
+      }
+      const time = new Date();
+      const screenshotTitle =
+        `${time.getFullYear()}-${time.getMonth()}-${time.getDay()}-${time.getUTCHours()}h-${time.getUTCMinutes()}m-${time.getUTCSeconds()}s-${
+          test.parent
+        }-${test.title}`.replace(/[ |/]/g, '-');
+      browser.saveScreenshot(`./.screenshots/${screenshotTitle}.png`);
+    }
+  },
 
   /**
    * Hook that gets executed after the suite has ended