+++ libs/shared/utils/src/lib/check-is-nx-workspace.ts
@@ -3,6 +3,8 @@ import {
   fileExists,
   readAndCacheJsonFile,
 } from '@nx-console/shared/file-system';
+import { workspaceDependencyPath } from '@nx-console/shared/npm';
+import { coerce } from 'semver';
 
 export async function checkIsNxWorkspace(
   workspacePath: string
@@ -10,8 +12,22 @@ export async function checkIsNxWorkspace(
   let isNxWorkspace = await fileExists(join(workspacePath, 'nx.json'));
 
   if (!isNxWorkspace) {
-    const lerna = await readAndCacheJsonFile('lerna.json', workspacePath);
-    isNxWorkspace = lerna.json.useNx ?? false;
+    const lernaPath = await workspaceDependencyPath(workspacePath, 'lerna');
+    if (!lernaPath) {
+      return false;
+    }
+
+    const lernaPackageJson = await readAndCacheJsonFile(
+      join(lernaPath, 'package.json')
+    );
+    const lernaVersion = coerce(lernaPackageJson.json.version);
+
+    if (lernaVersion?.major ?? 0 >= 6) {
+      isNxWorkspace = true;
+    } else {
+      const lerna = await readAndCacheJsonFile('lerna.json', workspacePath);
+      isNxWorkspace = lerna.json.useNx ?? false;
+    }
   }
 
   return isNxWorkspace;